configure building nested schema dynamically
- found that I can build nested schema more easily by 2 steps:
1- initiate a new object from the "form" item, not the "field" item
2- pass "shape" as the second parameter to configValidation function to control the scope

- Basic Idea: configValidation is building shape in a specific scope of
the shape passed to as a parameter.
To explain it more, here is an example of how it should and actually behave:

- if the item is "form": ( it would be a nested form to the main shape )
- create an empty newFormShape to store form children's schemas
- add fieldName variable to be used as key of form object in its parent shape
-- if it is not nested in another nested form ( no "." in its name ): set fieldName to the current item name
-- if it is a nested form to another nested form (has "." in its name ):
split name by "." and get the last item to be the fieldName
- recall configValidation with parameters:
-- current form children
-- newFormShape ( which is empty so far )
so we can work at the current form level ( adding fields to its own shape ), not the parent level
- add new schema with the newFormShape ( returned form configValidation )
to parent shape with fieldName as the key name

- if the item is "field":
-- if it's not nested ( no "." in its name ): added to current shape.
it will be the main shape since no nesting yet.
-- if it is nested ( has "." in its name ):disappointed:
In this case, the shape is already for the nested form and we are
working on the right scope, so split the name by "."
and get the last item to be field key name in the current shape

- it worth mentioning that using the "." separated naming convention for fields is now as an indicator and
not used for building nested object but I think for better description
I will keep using it